requirements.txt
aniso8601==9.0.1
attrs==21.2.0
certifi==2021.10.8
cffi==1.15.1
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.5
Flask==2.0.2
flask-restx==0.5.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==5.0.0
itsdangerous==2.0.1
Jinja2==3.0.2
jsonschema==4.1.2
MarkupSafe==2.0.1
marshmallow==3.14.0
pycparser==2.21
PyJWT==2.5.0
pyrsistent==0.18.0
pytz==2021.3
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.26
typing-extensions==4.4.0
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.9.0
